Create PROGRAM 6
Design, Develop and Implement a menu driven Program in C for the following
operations on Doubly Linked List (DLL) of Professor Data with the fields: ID, Name, Branch,
Area of specialization
a. Create a DLL stack of N Professor’s Data.
b. Create a DLL queue of N Professor’s Data
Display the status of DLL and count the number of nodes in it.